WIP on refactor
