+++ packages/scaffolder/src/run.ts
@@ -22,17 +22,17 @@ const welcome = (dryRun: boolean) => {
 export const run = async (argv: string[] | undefined, dryRun = false) => {
   welcome(dryRun);
 
-  const store = initializeFeatureStore();
+  const store = await initializeFeatureStore();
 
-  throw new Error('asdad');
+  // throw new Error('asdad');
 
-  const features = await getFeatures();
+  // const features = await getFeatures();
 
-  if (!features.length) {
-    handleError('No features found to scaffold.\n\nEnsure that your configuration isn\'t inadvertently overriding the built-in sources included with the scaffolder.');
-  }
+  // if (!features.length) {
+  //   handleError('No features found to scaffold.\n\nEnsure that your configuration isn\'t inadvertently overriding the built-in sources included with the scaffolder.');
+  // }
 
-  logger().debug(`Found ${features.length} features to scaffold: ${JSON.stringify(features, null, 2)}`);
+  // logger().debug(`Found ${features.length} features to scaffold: ${JSON.stringify(features, null, 2)}`);
 
   // const feature = await resolveFeature(features, argv ? argv[0] : undefined);
